refactor utilities + error function
